[feat] #305 - 불필요한 코드 제거

+++ data/src/main/java/com/ivyclub/data/image/ImageManager.kt
@@ -20,16 +20,13 @@ object ImageManager {
         }
     }
 
+    // 기존 파일 삭제하고 새로운 bitmap list로 다시 파일들 생성
     fun savePlanBitmap(bitmapList: List<Bitmap>, planId: String) {
         runCatching {
             val folderPath = "${ImageType.PLAN_IMAGE.filePath}${planId}/"
             val file = File(folderPath)
-            if (file.exists()) {
-                file.deleteRecursively()
-                file.mkdirs()
-            } else {
-                file.mkdirs()
-            }
+            if (file.exists()) file.deleteRecursively()
+            file.mkdirs()
             bitmapList.forEachIndexed { index, bitmap ->
                 val tempFile = File(folderPath, "${index}.jpg").apply {
                     createNewFile()
